Add final version of project

+++ mes_jakub_krech/Node.cpp
@@ -1,7 +1,7 @@
 #include "Node.h"
 #include <iostream>
 
-Node::Node(double xx, double yy, bool boundary_condition, int initial_temperature) :x{ xx }, y{ yy }, BC{ boundary_condition }, temperature{initial_temperature}
+Node::Node(double xx, double yy, bool boundary_condition, double initial_temperature) :x{ xx }, y{ yy }, BC{ boundary_condition }, temperature{initial_temperature}
 {
 	id = global_id++;
 }
